+++ shuup/admin/modules/currencies/views/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# This file is part of Shuup.
+#
+# Copyright (c) 2012-2016, Shoop Commerce Ltd. All rights reserved.
+#
+# This source code is licensed under the AGPLv3 license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .edit import CurrencyEditView
+from .list import CurrencyListView
+
+__all__ = [
+    "CurrencyEditView",
+    "CurrencyListView",
+]